refactor(be): use interface instead of type
object type은 interface가 더 적합하다
https://stackoverflow.com/questions/37233735/interfaces-vs-types-in-typescript
