+++ backend/src/auth/interface/authenticated-user.interface.ts
@@ -1,0 +1,5 @@
+export interface AuthenticatedUser {
+  id: number
+  username: string
+  role: string
+}
