+++ backend/src/auth/interface/jwt.interface.ts
@@ -1,0 +1,15 @@
+export interface JwtPayload {
+  userId: number
+  username: string
+}
+
+export interface JwtObject extends JwtPayload {
+  iat: number
+  exp: number
+  iss: string
+}
+
+export interface JwtTokens {
+  accessToken: string
+  refreshToken: string
+}
