+++ backend/src/auth/strategy/jwt.strategy.ts
@@ -4,7 +4,8 @@ import { PassportStrategy } from '@nestjs/passport'
 import { ExtractJwt } from 'passport-jwt'
 import { Strategy } from 'passport-jwt'
 import { UserService } from 'src/user/user.service'
-import { JwtObject, AuthenticatedUser } from '../type/jwt.type'
+import { JwtObject } from '../interface/jwt.interface'
+import { AuthenticatedUser } from '../interface/authenticated-user.interface'
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
